pyproject.toml
[project]
name = "multisetp-form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.2",
    "django-crispy-forms>=2.4",
    "django-formtools>=2.5.1",
]
